added test on real case
